FIXME zero_grad() should be called on opt not model
This is because opt is the optimizer object that is responsible for updating the model parameters based on the gradients computed by the backward pass, and therefore it should be the one to call zero_grad() on.

Calling model.zero_grad() would only reset the gradients of the model parameters, but the optimizer would still retain the previous values and update the parameters with incorrect gradients, leading to incorrect parameter updates and suboptimal performance.